+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/activities/WidgetConfigurationActivity.java
@@ -19,6 +19,7 @@ import java.util.List;
 import java.util.Map;
 
 import nodomain.freeyourgadget.gadgetbridge.GBApplication;
+import nodomain.freeyourgadget.gadgetbridge.R;
 import nodomain.freeyourgadget.gadgetbridge.Widget;
 import nodomain.freeyourgadget.gadgetbridge.database.DBHandler;
 import nodomain.freeyourgadget.gadgetbridge.database.DBHelper;
@@ -58,7 +59,7 @@ public class WidgetConfigurationActivity extends Activity {
         }
 
         AlertDialog.Builder builder = new AlertDialog.Builder(WidgetConfigurationActivity.this);
-        builder.setTitle("Select device");
+        builder.setTitle(R.string.widget_settings_select_device_title);
 
         allDevices = getAllDevices(getApplicationContext());
 
@@ -118,12 +119,11 @@ public class WidgetConfigurationActivity extends Activity {
                 Device dbDevice = DBHelper.findDevice(device, daoSession);
                 int icon = device.isInitialized() ? device.getType().getIcon() : device.getType().getDisabledIcon();
                 if (dbDevice != null && coordinator != null
-                        && coordinator.supportsActivityTracks()
+                        && (coordinator.supportsActivityDataFetching() || coordinator.supportsActivityTracking())
                         && !newMap.containsKey(device.getAliasOrName())) {
                     newMap.put(device.getAliasOrName(), new Pair(device.getAddress(), icon));
                 }
             }
-
         } catch (Exception e) {
             LOG.error("Error getting list of all devices: " + e);
         }